# Allow using scalac plugins that emit additional compilation results