--- a/src/python/pants/backend/scala/compile/scalac.py
+++ b/src/python/pants/backend/scala/compile/scalac.py
@@ -25,8 +25,17 @@ from pants.backend.scala.util_rules.versions import (
     ScalaArtifactsForVersionResult,
     ScalaVersion,
 )
-from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
-from pants.engine.fs import EMPTY_DIGEST, Digest, MergeDigests
+from pants.core.util_rules.source_files import SourceFilesRequest
+from pants.core.util_rules.stripped_source_files import StrippedSourceFiles
+from pants.engine.fs import (
+    EMPTY_DIGEST,
+    CreateDigest,
+    Digest,
+    Directory,
+    MergeDigests,
+    RemovePrefix,
+    Snapshot,
+)
 from pants.engine.process import FallibleProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import CoarsenedTarget, SourcesField
@@ -41,9 +50,12 @@ from pants.jvm.compile import (
     FallibleClasspathEntry,
 )
 from pants.jvm.compile import rules as jvm_compile_rules
+from pants.jvm.jar_tool import jar_tool
+from pants.jvm.jar_tool.jar_tool import JarToolRequest
 from pants.jvm.jdk_rules import JdkEnvironment, JdkRequest, JvmProcess
 from pants.jvm.resolve.common import ArtifactRequirements
 from pants.jvm.resolve.coursier_fetch import ToolClasspath, ToolClasspathRequest
+from pants.jvm.strip_jar import strip_jar
 from pants.jvm.strip_jar.strip_jar import StripJarRequest
 from pants.jvm.subsystems import JvmSubsystem
 from pants.jvm.target_types import NO_MAIN_CLASS
@@ -98,7 +110,9 @@ async def compile_scala_source(
         component_members_with_sources,
         await MultiGet(
             Get(
-                SourceFiles,
+                # Some Scalac plugins (i.e. SemanticDB) require us to use stripped source files so the plugin
+                # would emit compilation output that correlates with the appropiate paths in the input files.
+                StrippedSourceFiles,
                 SourceFilesRequest(
                     (t.get(SourcesField),),
                     for_sources_types=(ScalaSourceField, JavaSourceField),
@@ -175,6 +189,10 @@ async def compile_scala_source(
     classpath_arg = ":".join(user_classpath.immutable_inputs_args(prefix=usercp))
 
     output_file = compute_output_jar_filename(request.component)
+    compilation_output_dir = "__out"
+    compilation_empty_dir = await Get(Digest, CreateDigest([Directory(compilation_output_dir)]))
+    merged_digest = await Get(Digest, MergeDigests([sources_digest, compilation_empty_dir]))
+
     process_result = await Get(
         FallibleProcessResult,
         JvmProcess(
@@ -187,12 +205,8 @@ async def compile_scala_source(
                 *local_plugins.args(local_scalac_plugins_relpath),
                 *(("-classpath", classpath_arg) if classpath_arg else ()),
                 *scalac.args,
-                # NB: We set a non-existent main-class so that using `-d` produces a `jar` manifest
-                # with stable content.
-                "-Xmain-class",
-                NO_MAIN_CLASS,
                 "-d",
-                output_file,
+                compilation_output_dir,
                 *sorted(
                     chain.from_iterable(
                         sources.snapshot.files
@@ -200,17 +214,32 @@ async def compile_scala_source(
                     )
                 ),
             ],
-            input_digest=sources_digest,
+            input_digest=merged_digest,
             extra_immutable_input_digests=extra_immutable_input_digests,
             extra_nailgun_keys=extra_nailgun_keys,
-            output_files=(output_file,),
+            output_directories=(compilation_output_dir,),
             description=f"Compile {request.component} with scalac",
             level=LogLevel.DEBUG,
         ),
     )
     output: ClasspathEntry | None = None
     if process_result.exit_code == 0:
-        output_digest = process_result.output_digest
+        compilation_results = await Get(
+            Snapshot, RemovePrefix(process_result.output_digest, compilation_output_dir)
+        )
+
+        output_digest = await Get(
+            Digest,
+            JarToolRequest(
+                jar_name=output_file,
+                digest=compilation_results.digest,
+                # NB: We set a non-existent main-class so we produce a `jar` manifest
+                # with stable content.
+                main_class=NO_MAIN_CLASS,
+                files=compilation_results.files,
+            ),
+        )
+
         if jvm.reproducible_jars:
             output_digest = await Get(
                 Digest, StripJarRequest(digest=output_digest, filenames=(output_file,))
@@ -251,5 +280,7 @@ def rules():
         *scala_artifact_rules(),
         *scalac_plugins_rules(),
         *versions.rules(),
+        *jar_tool.rules(),
+        *strip_jar.rules(),
         UnionRule(ClasspathEntryRequest, CompileScalaSourceRequest),
     ]