--- a/src/python/pants/backend/scala/compile/scalac.py
+++ b/src/python/pants/backend/scala/compile/scalac.py
@@ -27,16 +27,9 @@ from pants.backend.scala.util_rules.versions import (
 )
 from pants.core.util_rules.source_files import SourceFilesRequest
 from pants.core.util_rules.stripped_source_files import StrippedSourceFiles
-from pants.engine.fs import (
-    EMPTY_DIGEST,
-    CreateDigest,
-    Digest,
-    Directory,
-    MergeDigests,
-    RemovePrefix,
-    Snapshot,
-)
-from pants.engine.process import FallibleProcessResult
+from pants.core.util_rules.system_binaries import BashBinary, ZipBinary
+from pants.engine.fs import EMPTY_DIGEST, CreateDigest, Digest, Directory, MergeDigests
+from pants.engine.process import FallibleProcessResult, Process, ProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import CoarsenedTarget, SourcesField
 from pants.engine.unions import UnionRule
@@ -50,15 +43,12 @@ from pants.jvm.compile import (
     FallibleClasspathEntry,
 )
 from pants.jvm.compile import rules as jvm_compile_rules
-from pants.jvm.jar_tool import jar_tool
-from pants.jvm.jar_tool.jar_tool import JarToolRequest
 from pants.jvm.jdk_rules import JdkEnvironment, JdkRequest, JvmProcess
 from pants.jvm.resolve.common import ArtifactRequirements
 from pants.jvm.resolve.coursier_fetch import ToolClasspath, ToolClasspathRequest
 from pants.jvm.strip_jar import strip_jar
 from pants.jvm.strip_jar.strip_jar import StripJarRequest
 from pants.jvm.subsystems import JvmSubsystem
-from pants.jvm.target_types import NO_MAIN_CLASS
 from pants.util.logging import LogLevel
 
 logger = logging.getLogger(__name__)
@@ -84,6 +74,8 @@ async def compile_scala_source(
     scala: ScalaSubsystem,
     jvm: JvmSubsystem,
     scalac: Scalac,
+    bash: BashBinary,
+    zip_binary: ZipBinary,
     request: CompileScalaSourceRequest,
 ) -> FallibleClasspathEntry:
     # Request classpath entries for our direct dependencies.
@@ -193,7 +185,7 @@ async def compile_scala_source(
     compilation_empty_dir = await Get(Digest, CreateDigest([Directory(compilation_output_dir)]))
     merged_digest = await Get(Digest, MergeDigests([sources_digest, compilation_empty_dir]))
 
-    process_result = await Get(
+    compile_result = await Get(
         FallibleProcessResult,
         JvmProcess(
             jdk=jdk,
@@ -223,22 +215,34 @@ async def compile_scala_source(
         ),
     )
     output: ClasspathEntry | None = None
-    if process_result.exit_code == 0:
-        compilation_results = await Get(
-            Snapshot, RemovePrefix(process_result.output_digest, compilation_output_dir)
-        )
-
-        output_digest = await Get(
-            Digest,
-            JarToolRequest(
-                jar_name=output_file,
-                digest=compilation_results.digest,
-                # NB: We set a non-existent main-class so we produce a `jar` manifest
-                # with stable content.
-                main_class=NO_MAIN_CLASS,
-                files=compilation_results.files,
+    if compile_result.exit_code == 0:
+        # We package the outputs into a JAR file in a similar way as how it's
+        # done in the `javac.py` implementation
+        jar_result = await Get(
+            ProcessResult,
+            Process(
+                argv=[
+                    bash.path,
+                    "-c",
+                    " ".join(
+                        [
+                            "cd",
+                            compilation_output_dir,
+                            ";",
+                            zip_binary.path,
+                            "-r",
+                            f"../{output_file}",
+                            ".",
+                        ]
+                    ),
+                ],
+                input_digest=compile_result.output_digest,
+                output_files=(output_file,),
+                description=f"Capture outputs of {request.component} for scalac",
+                level=LogLevel.TRACE,
             ),
         )
+        output_digest = jar_result.output_digest
 
         if jvm.reproducible_jars:
             output_digest = await Get(
@@ -249,7 +253,7 @@ async def compile_scala_source(
 
     return FallibleClasspathEntry.from_fallible_process_result(
         str(request.component),
-        process_result,
+        compile_result,
         output,
     )
 
@@ -280,7 +284,6 @@ def rules():
         *scala_artifact_rules(),
         *scalac_plugins_rules(),
         *versions.rules(),
-        *jar_tool.rules(),
         *strip_jar.rules(),
         UnionRule(ClasspathEntryRequest, CompileScalaSourceRequest),
     ]