--- a/src/python/pants/jvm/compile_test.py
+++ b/src/python/pants/jvm/compile_test.py
@@ -311,7 +311,6 @@ def test_compile_mixed(
     )
 
     assert rendered_classpath.content[".Example.scala.main.scalac.jar"] == {
-        "META-INF/MANIFEST.MF",
         "org/pantsbuild/example/Main$.class",
         "org/pantsbuild/example/Main.class",
     }