… src/python/pants/backend/scala/compile/acyclic.test.lock
… src/python/pants/backend/scala/compile/joda-time.test.lock
… src/python/pants/backend/scala/compile/multiple-scalac-plugins.test.lock
… src/python/pants/backend/scala/compile/scala-library-2.12.test.lock
… src/python/pants/backend/scala/compile/scala-library-2.13.test.lock