Fixes #20356.

When using `scalac` plugins that emit additional compilation results, `scalac` changes the behavior of the `-d` flag and instead of producing a JAR file, it produces a directory containing all the results of the different compilation stages.

The fix consists on forcing `scalac` to always emit results into a directory by passing an empty dir as destination and then using ~the internal `jartool`~ `zip` to package everything into a single JAR.

Sliced from #20328.