@alonsodomin 2023-12-31
It's a perfectly valid concern and to be fair, I haven't really measured it. My expectation is that on first run it will defo have a performance penalty as the `jartool` needs to be built and its dependencies will need to be fetch to so.

On normal use I haven't noticed a big impact when building Scala code and I guess that's mostly down to Pant's caching. Obviously this totally subjective.