@tdyas 2023-12-31
Maybe collect some timing with and without the change for a Scala repo (e.g., example-jvm?)?

i recall @Eric-Arellano has used hyperfine in the past to collect timing runs on Pants.