@Eric-Arellano 2024-01-01
> i recall @Eric-Arellano has used hyperfine in the past to collect timing runs on Pants.

https://www.pantsbuild.org/docs/contributions-debugging