@alonsodomin 2024-01-04
Sorry for taking long to come back to this, but manage to come around and make some measurements using `hyperfine` comparing performance of `pants check ::` in the `example-jvm` repository using the `main` branch and the one for this PR (3 warm runs and 5 measurement runs).

## Main branch, local cache disabled

```shell
> hyperfine --warmup=3 --runs=5 './pants_from_sources --no-pantsd --no-local-cache check ::'

Benchmark 1: ./pants_from_sources --no-pantsd --no-local-cache check ::
  Time (mean ± σ):     18.512 s ±  0.118 s    [User: 16.493 s, System: 5.196 s]
  Range (min … max):   18.366 s … 18.637 s    5 runs
```

## Main branch, local cache enabled

```shell
> hyperfine --warmup=3 --runs=5 './pants_from_sources --no-pantsd --no-local-cache check ::'

Benchmark 1: ./pants_from_sources --no-pantsd --no-local-cache check ::
  Time (mean ± σ):     20.441 s ±  0.254 s    [User: 19.295 s, System: 5.539 s]
  Range (min … max):   20.084 s … 20.707 s    5 runs
```

## PR branch, local cache disabled

```shell
> hyperfine --warmup=3 --runs=5 './pants_from_sources --no-pantsd check ::'

Benchmark 1: ./pants_from_sources --no-pantsd check ::
  Time (mean ± σ):      8.091 s ±  0.150 s    [User: 3.794 s, System: 0.433 s]
  Range (min … max):    7.864 s …  8.260 s    5 runs
```

## PR branch, local cache enabled

```shell
> hyperfine --warmup=3 --runs=5 './pants_from_sources --no-pantsd check ::'

Benchmark 1: ./pants_from_sources --no-pantsd check ::
  Time (mean ± σ):      8.874 s ±  0.616 s    [User: 3.897 s, System: 0.488 s]
  Range (min … max):    8.149 s …  9.697 s    5 runs
```

So there is a clear impact of about 2 seconds when running without the local cache. The difference becomes much less noticeable when the cache is enabled with a range from 0.5 to 1.5 seconds in the best and worse scenario.

Wondering how this would scale up in a very large repository though (not having a large enough one at hand right now), but if we wouldn't like to assume this penalty we would need to find another way to obtain the `ClasspathEntry`. Options that come to mind are either packaging the JAR using the `zip` system binary instead or basically don't package the results at all and use an unpacked `Digest` with all the results expanded.