@alonsodomin 2024-01-04
just tested the same but using `zip` to package the JAR file instead of `jartool` (in the same way as it is done in `javac.py`) and the measured times we basically indistinguishable between main branch and this change, so this alternative seems a better option performance-wise.